fighting mypy, round #2345235923523099836

+++ rerun_py/rerun_sdk/rerun2/components/color_ext.py
@@ -38,7 +38,7 @@ class ColorArrayExt:
             elif array.dtype.type == np.uint32:
                 array = np.asarray(data).flatten()
             else:
-                array = np.asarray(data).reshape((-1, 4))
+                array = np.asarray(data, dtype=np.uint8).reshape((-1, 4))
                 array = u8_array_to_rgba(array)
 
         return arrow().wrap_array(pa.array(array, type=arrow().storage_type))
